+++ src/main/java/business/NewsSources/ZingNews.java
@@ -1,10 +1,10 @@
-package BusinessLayer.NewsSources;
+package business.NewsSources;
 
-import BusinessLayer.Helper.CATEGORY;
-import BusinessLayer.Helper.CSS;
-import BusinessLayer.Helper.ParseLocalDateTime;
-import BusinessLayer.Sanitizer.HtmlSanitizer;
-import BusinessLayer.Sanitizer.ZingNewsSanitizer;
+import business.Helper.CATEGORY;
+import business.Helper.CSS;
+import business.Helper.LocalDateTimeParser;
+import business.Sanitizer.HtmlSanitizer;
+import business.Sanitizer.ZingNewsSanitizer;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
@@ -60,7 +60,7 @@ public class ZingNews extends NewsOutlet{
         if (dateTimeStr.isEmpty() || dateTimeStr.isBlank()){
             return LocalDateTime.now();
         }
-        return ParseLocalDateTime.parse(dateTimeStr);
+        return LocalDateTimeParser.parse(dateTimeStr);
     }
 
     @Override
