+++ src/main/java/business/Sanitizer/ThanhNienSanitizer.java
@@ -1,11 +1,11 @@
-package BusinessLayer.Sanitizer;
+package business.Sanitizer;
 
-import BusinessLayer.Helper.CSS;
+import business.Helper.CSS;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Element;
 import org.jsoup.safety.Safelist;
 
-public class TuoiTreSanitizer extends HtmlSanitizer {
+public class ThanhNienSanitizer extends HtmlSanitizer {
     @Override
     protected Element sanitizeNonTitleTag(Element e, String type) {
         Safelist safelist; // modify this safe list according to the type
@@ -17,7 +17,6 @@ public class TuoiTreSanitizer extends HtmlSanitizer {
                 safelist = Safelist.basic();
                 cleanHtml = Jsoup.clean(e.html(), safelist);
                 newHtmlElement = new Element("p").html(cleanHtml);
-
                 return newHtmlElement;
             case CSS.MAIN_CONTENT:
 
@@ -26,7 +25,6 @@ public class TuoiTreSanitizer extends HtmlSanitizer {
 
                 break;
         }
-
         return e;
     }
 }
