+++ src/main/java/business/Sanitizer/TuoiTreSanitizer.java
@@ -1,11 +1,11 @@
-package BusinessLayer.Sanitizer;
+package business.Sanitizer;
 
-import BusinessLayer.Helper.CSS;
+import business.Helper.CSS;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Element;
 import org.jsoup.safety.Safelist;
 
-public class ZingNewsSanitizer extends HtmlSanitizer {
+public class TuoiTreSanitizer extends HtmlSanitizer {
     @Override
     protected Element sanitizeNonTitleTag(Element e, String type) {
         Safelist safelist; // modify this safe list according to the type
@@ -20,14 +20,13 @@ public class ZingNewsSanitizer extends HtmlSanitizer {
 
                 return newHtmlElement;
             case CSS.MAIN_CONTENT:
-                safelist = Safelist.relaxed();
-
 
                 break;
             default:
 
                 break;
         }
+
         return e;
     }
 }
